replace: carriages data type with unsignedTinyInteger

+++ database/migrations/2023_10_16_151053_create_trains_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('arrival_station', 50);
             $table->time('departure', $precision = 0);
             $table->time('arrival', $precision = 0);
-            $table->smallInteger('carriages');
+            $table->unsignedTinyInteger('carriages');
             $table->boolean('delayed');
             $table->boolean('canceled');
             $table->timestamps();
